Added Test data in pgAdmin

+++ src/main/java/com/springbootpractice/quizapp/dao/QuestionDao.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface QuestionDao extends JpaRepository<Question, Integer> {
+public interface QuestionDao extends JpaRepository<Question, Long> {
     List<Question> findByCategory(String category);
 }
